+++ pkg/option/option_test.go
@@ -1,13 +1,13 @@
 package option_test
 
 import (
-	"fmt"
 	"net/url"
 	"reflect"
 	"testing"
 	"time"
 
 	"github.com/maargenton/go-errors"
+	"github.com/maargenton/go-testpredicate/pkg/bdd"
 	"github.com/maargenton/go-testpredicate/pkg/require"
 	"github.com/maargenton/go-testpredicate/pkg/verify"
 
@@ -20,7 +20,7 @@ import (
 // ---------------------------------------------------------------------------
 
 func TestOptionName(t *testing.T) {
-	t.Run("Given a struct with opts tags", func(t *testing.T) {
+	bdd.Given(t, "a struct with opts tags", func(t *bdd.T) {
 		type options struct {
 			Period time.Duration `opts:"-d,--duration"`
 			Count  int           `opts:"-c"`
@@ -34,51 +34,51 @@ func TestOptionName(t *testing.T) {
 		require.That(t, err).IsNil()
 		require.That(t, optionSet).IsNotNil()
 
-		t.Run("when getting the name of a field with long flag", func(t *testing.T) {
+		t.When("getting the name of a field with long flag", func(t *bdd.T) {
 			var option = optionSet.GetOption("duration")
 			require.That(t, option).IsNotNil()
 
-			t.Run("the name includes the long flag", func(t *testing.T) {
+			t.Then("the name includes the long flag", func(t *bdd.T) {
 				require.That(t, option.Name()).Eq("--duration")
 			})
 		})
 
-		t.Run("when getting the name of a field with only short flag", func(t *testing.T) {
+		t.When("getting the name of a field with only short flag", func(t *bdd.T) {
 			var option = optionSet.GetOption("c")
 			require.That(t, option).IsNotNil()
 
-			t.Run("the name includes the short flag", func(t *testing.T) {
+			t.Then("the name includes the short flag", func(t *bdd.T) {
 				require.That(t, option.Name()).Eq("-c")
 			})
 		})
 
-		t.Run("when getting the name of a positional argument field", func(t *testing.T) {
+		t.When("getting the name of a positional argument field", func(t *bdd.T) {
 			var option = optionSet.Positional[0]
 			require.That(t, option).IsNotNil()
 
-			t.Run("the name includes the value name", func(t *testing.T) {
+			t.Then("the name includes the value name", func(t *bdd.T) {
 				require.That(t, option.Name()).Eq("<name>")
 			})
 		})
-		t.Run("when getting the name of a positional argument field", func(t *testing.T) {
+		t.When("getting the name of a positional argument field", func(t *bdd.T) {
 			var option = optionSet.Positional[1]
 			require.That(t, option).IsNotNil()
 
-			t.Run("the name includes the argument position", func(t *testing.T) {
+			t.Then("the name includes the argument position", func(t *bdd.T) {
 				require.That(t, option.Name()).Eq("<arg2>")
 			})
 		})
-		t.Run("when getting the name of an extra arguments field", func(t *testing.T) {
+		t.When("getting the name of an extra arguments field", func(t *bdd.T) {
 			var option = optionSet.Args
 			require.That(t, option).IsNotNil()
 
-			t.Run("the name includes the value name", func(t *testing.T) {
+			t.Then("the name includes the value name", func(t *bdd.T) {
 				require.That(t, option.Name()).Eq("<inputs>...")
 			})
 		})
 	})
 
-	t.Run("Given another struct with opts tags", func(t *testing.T) {
+	bdd.Given(t, "another struct with opts tags", func(t *bdd.T) {
 		type options struct {
 			Inputs []string `opts:"args"`
 		}
@@ -88,11 +88,11 @@ func TestOptionName(t *testing.T) {
 		require.That(t, err).IsNil()
 		require.That(t, optionSet).IsNotNil()
 
-		t.Run("when getting the name of an extra arguments field", func(t *testing.T) {
+		t.When("getting the name of an extra arguments field", func(t *bdd.T) {
 			var option = optionSet.Args
 			require.That(t, option).IsNotNil()
 
-			t.Run("the name includes default name and ellipsis", func(t *testing.T) {
+			t.Then("the name includes default name and ellipsis", func(t *bdd.T) {
 				require.That(t, option.Name()).Eq("<args>...")
 			})
 		})
@@ -193,9 +193,9 @@ func TestOptionDescriptionUsage(t *testing.T) {
 	}
 
 	for _, tc := range tcs {
-		t.Run(fmt.Sprintf("Given %v", tc.name), func(t *testing.T) {
-			t.Run("when calling Usage()", func(t *testing.T) {
-				t.Run("then it returns formatted usage", func(t *testing.T) {
+		bdd.Given(t, tc.name, func(t *bdd.T) {
+			t.When("calling Usage()", func(t *bdd.T) {
+				t.Then("it returns formatted usage", func(t *bdd.T) {
 					require.That(t, tc.opt.GetUsage()).Field("Option").Eq(tc.usage)
 				})
 			})
@@ -241,9 +241,9 @@ func TestOptionDescriptionDescription(t *testing.T) {
 	}
 
 	for _, tc := range tcs {
-		t.Run(fmt.Sprintf("Given %v", tc.name), func(t *testing.T) {
-			t.Run("when calling Usage()", func(t *testing.T) {
-				t.Run("then it returns formatted description", func(t *testing.T) {
+		bdd.Given(t, tc.name, func(t *bdd.T) {
+			t.When("calling Usage()", func(t *bdd.T) {
+				t.Then("it returns formatted description", func(t *bdd.T) {
 					require.That(t, tc.opt.GetUsage()).Field("Description").Eq(tc.desc)
 				})
 			})
@@ -256,7 +256,7 @@ func TestOptionDescriptionDescription(t *testing.T) {
 // ---------------------------------------------------------------------------
 
 func TestOption_SetBool(t *testing.T) {
-	t.Run("Given a struct with bool options", func(t *testing.T) {
+	bdd.Given(t, "a struct with bool options", func(t *bdd.T) {
 		args := struct {
 			A   bool  `opts:"-a,--aaa"`
 			Ptr *bool `opts:"-p,--ppp"`
@@ -265,25 +265,25 @@ func TestOption_SetBool(t *testing.T) {
 		optionSet, err := option.NewOptionSet(&args)
 		require.That(t, err).IsNil()
 
-		t.Run("when calling Set() on a bool field", func(t *testing.T) {
+		t.When("calling Set() on a bool field", func(t *bdd.T) {
 			optionSet.GetOption("a").SetBool()
 
-			t.Run("then the field is set to true", func(t *testing.T) {
+			t.Then("the field is set to true", func(t *bdd.T) {
 				require.That(t, args.A).IsTrue()
 			})
 		})
 
-		t.Run("when calling Set() on a bool pointer field", func(t *testing.T) {
+		t.When("calling Set() on a bool pointer field", func(t *bdd.T) {
 			optionSet.GetOption("p").SetBool()
 
-			t.Run("then the field is set to point to a true value", func(t *testing.T) {
+			t.Then("the field is set to point to a true value", func(t *bdd.T) {
 				require.That(t, args.Ptr).IsNotNil()
 				require.That(t, *args.Ptr).IsTrue()
 			})
 		})
 
-		t.Run("when calling Set() on a non-bool field", func(t *testing.T) {
-			t.Run("then it panics", func(t *testing.T) {
+		t.When("calling Set() on a non-bool field", func(t *bdd.T) {
+			t.Then("it panics", func(t *bdd.T) {
 
 				require.That(t, func() {
 					optionSet.GetOption("n").SetBool()
@@ -300,7 +300,7 @@ func TestOption_SetBool(t *testing.T) {
 // ---------------------------------------------------------------------------
 
 func TestOption_SetValue(t *testing.T) {
-	t.Run("Given a struct with `opts` tags", func(t *testing.T) {
+	bdd.Given(t, "a struct with `opts` tags", func(t *bdd.T) {
 
 		args := struct {
 			Period    time.Duration   `opts:"-p,--period"`
@@ -311,71 +311,71 @@ func TestOption_SetValue(t *testing.T) {
 		require.That(t, err).IsNil()
 		require.That(t, optionSet).IsNotNil()
 
-		t.Run("when setting the value of a scalar field", func(t *testing.T) {
+		t.When("setting the value of a scalar field", func(t *bdd.T) {
 			option := optionSet.GetOption("period")
 			require.That(t, option).IsNotNil()
 
-			t.Run("with a valid value", func(t *testing.T) {
+			t.Run("with a valid value", func(t *bdd.T) {
 				args.Period = 0 * time.Second
 				err := option.SetValue("5m30s")
 				expected := 5*time.Minute + 30*time.Second
 
-				t.Run("then the field is set accordingly", func(t *testing.T) {
+				t.Then("the field is set accordingly", func(t *bdd.T) {
 					require.That(t, err).IsNil()
 					require.That(t, args.Period).Eq(expected)
 				})
 			})
 
-			t.Run("with an invalid value", func(t *testing.T) {
+			t.Run("with an invalid value", func(t *bdd.T) {
 				args.Period = 0 * time.Second
 				err := option.SetValue("5m30p")
 
-				t.Run("then an error is returned and the value is not changed", func(t *testing.T) {
+				t.Then("an error is returned and the value is not changed", func(t *bdd.T) {
 					require.That(t, err).IsNotNil()
 					require.That(t, args.Period).Eq(0 * time.Second)
 				})
 			})
 		})
 
-		t.Run("when setting the value of a pointer field", func(t *testing.T) {
+		t.When("setting the value of a pointer field", func(t *bdd.T) {
 			option := optionSet.GetOption("duration")
 			require.That(t, option).IsNotNil()
 
-			t.Run("with a valid value", func(t *testing.T) {
+			t.Run("with a valid value", func(t *bdd.T) {
 				args.Duration = nil
 				err := option.SetValue("5m30s")
 				expected := 5*time.Minute + 30*time.Second
 
-				t.Run("then the field is set accordingly", func(t *testing.T) {
+				t.Then("the field is set accordingly", func(t *bdd.T) {
 					require.That(t, err).IsNil()
 					require.That(t, args.Duration).IsNotNil()
 					require.That(t, *args.Duration).Eq(expected)
 				})
 			})
 
-			t.Run("with an invalid value", func(t *testing.T) {
+			t.Run("with an invalid value", func(t *bdd.T) {
 				args.Duration = nil
 				err := option.SetValue("5m30p")
 
-				t.Run("then an error is returned and the value is not changed", func(t *testing.T) {
+				t.Then("an error is returned and the value is not changed", func(t *bdd.T) {
 					require.That(t, err).IsNotNil()
 					require.That(t, args.Duration).IsNil()
 				})
 			})
 		})
 
-		t.Run("when setting the value of a slice field", func(t *testing.T) {
+		t.When("setting the value of a slice field", func(t *bdd.T) {
 			option := optionSet.GetOption("intervals")
 			require.That(t, option).IsNotNil()
 
-			t.Run("with a single valid value", func(t *testing.T) {
+			t.Run("with a single valid value", func(t *bdd.T) {
 				args.Intervals = nil
 				err := option.SetValue("5m30s")
 				expected := []time.Duration{
 					5*time.Minute + 30*time.Second,
 				}
 
-				t.Run("then the field is set accordingly", func(t *testing.T) {
+				t.Then("the field is set accordingly", func(t *bdd.T) {
 					require.That(t, err).IsNil()
 					require.That(t, args.Intervals).IsNotNil()
 					require.That(t, args.Intervals).Length().Eq(1)
@@ -383,7 +383,7 @@ func TestOption_SetValue(t *testing.T) {
 				})
 			})
 
-			t.Run("with multiple delimited values", func(t *testing.T) {
+			t.Run("with multiple delimited values", func(t *bdd.T) {
 				args.Intervals = nil
 				err1 := option.SetValue("1m,2m,3m")
 				err2 := option.SetValue("4m,5m")
@@ -395,14 +395,14 @@ func TestOption_SetValue(t *testing.T) {
 					5 * time.Minute,
 				}
 
-				t.Run("then all values are recorded", func(t *testing.T) {
+				t.Then("all values are recorded", func(t *bdd.T) {
 					require.That(t, err1).IsNil()
 					require.That(t, err2).IsNil()
 					require.That(t, args.Intervals).Eq(expected)
 				})
 			})
 
-			t.Run("with empty value", func(t *testing.T) {
+			t.Run("with empty value", func(t *bdd.T) {
 				args.Intervals = []time.Duration{
 					1 * time.Minute,
 					2 * time.Minute,
@@ -412,17 +412,17 @@ func TestOption_SetValue(t *testing.T) {
 				}
 				err := option.SetValue("")
 
-				t.Run("then all previous values are deleted", func(t *testing.T) {
+				t.Then("all previous values are deleted", func(t *bdd.T) {
 					require.That(t, err).IsNil()
 					require.That(t, args.Intervals).IsEmpty()
 				})
 			})
 
-			t.Run("with invalid value", func(t *testing.T) {
+			t.Run("with invalid value", func(t *bdd.T) {
 				args.Intervals = nil
 				err := option.SetValue("1m,2p,3m")
 
-				t.Run("then all values are recorded", func(t *testing.T) {
+				t.Then("all values are recorded", func(t *bdd.T) {
 					require.That(t, err).IsNotNil()
 					require.That(t, args.Intervals).IsEmpty()
 				})
@@ -454,4 +454,59 @@ func TestOption_SetValue_Ptr(t *testing.T) {
 	verify.That(t, opt.FieldType).Eq(urlType)
 	verify.That(t, opt.ValueType).Eq(urlType)
 	verify.That(t, opt.Type).Eq(option.Value)
+}
+
+// ---------------------------------------------------------------------------
+// Option.SetValue() -- slice type
+// ---------------------------------------------------------------------------
+
+func TestOption_SetValue_Slices(t *testing.T) {
+	bdd.Given(t, "an option set with a string-slice field", func(t *bdd.T) {
+		args := struct {
+			Default    []string `opts:"-v, --values, sep:\\,, env:DEFAULT"`
+			KeepSpaces []string `opts:"-v, --values, sep:\\,, keep-spaces, env:KEEP_SPACES"`
+			KeepEmpty  []string `opts:"-v, --values, sep:\\,, keep-empty, env:KEEP_EMPTY"`
+			KeepBoth   []string `opts:"-v, --values, sep:\\,, keep-spaces,keep-empty, env:KEEP_BOTH"`
+		}{}
+		opts, err := option.NewOptionSet(&args)
+		require.That(t, err).IsNil()
+		require.That(t, opts).IsNotNil()
+
+		t.When("using field with default options (trim spaces, discard empty)", func(t *bdd.T) {
+			opt := opts.Options[0]
+			opt.SetValue("aaa, bbb, ,")
+
+			t.Then("spaces are trimmed from the individual values", func(t *bdd.T) {
+				verify.That(t, args.Default).Length().Eq(2)
+				verify.That(t, args.Default).Eq([]string{"aaa", "bbb"})
+			})
+		})
+		t.When("using field with keep-spaces", func(t *bdd.T) {
+			opt := opts.Options[1]
+			opt.SetValue("aaa, bbb, ")
+
+			t.Then("spaces are preserved for the individual values", func(t *bdd.T) {
+				verify.That(t, args.KeepSpaces).Length().Eq(3)
+				verify.That(t, args.KeepSpaces).Eq([]string{"aaa", " bbb", " "})
+			})
+		})
+		t.When("using field with keep-empty", func(t *bdd.T) {
+			opt := opts.Options[2]
+			opt.SetValue("aaa, bbb,,")
+
+			t.Then("spaces are trimmed but empty values are preserved", func(t *bdd.T) {
+				verify.That(t, args.KeepEmpty).Length().Eq(3)
+				verify.That(t, args.KeepEmpty).Eq([]string{"aaa", "bbb", ""})
+			})
+		})
+		t.When("using field with keep-spaces and keep-empty", func(t *bdd.T) {
+			opt := opts.Options[3]
+			opt.SetValue("aaa, bbb,, ")
+
+			t.Then("all values are preserved with their spaces", func(t *bdd.T) {
+				verify.That(t, args.KeepBoth).Length().Eq(4)
+				verify.That(t, args.KeepBoth).Eq([]string{"aaa", " bbb", "", " "})
+			})
+		})
+	})
 }